backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.3
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy<2.0.0
scikit-surprise>=1.1.3
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
